initial MVC, calories subtract update
