+++ src/ai/generateAI.py
@@ -1,5 +1,6 @@
-import generateAIAction as g_act
-from generateAIAction import linkFoodIdToMealId
+import sys
+sys.path.insert(1, './')
+import src.dao.generateAIDAO as g_ai_dao
 
 mealNameBase = ["Breakfast","Lunch","Dinner","Snack"]
 
@@ -11,7 +12,7 @@ def generatePlanAI(planName: str, numMeal:int ):
         numMeal (int): The number of meals that will be generated by our ai
 
     Returns:
-        mealIDs (list): List of meal IDs that is generated by the plan
+        (void)
     """
     # FIXME: hardcode nummeal if its not between 0-3
     if numMeal > 4 or numMeal <= 0: numMeal = 3
@@ -20,13 +21,12 @@ def generatePlanAI(planName: str, numMeal:int ):
     planCalories = 5000
 
     # TODO: Generate a plan
-    planID = g_act.createNewPlanSQL(planName, planCalories)
+    planID = g_ai_dao.createNewPlanSQL(planName, planCalories)
 
     # TODO: Link plan with all of the meals
-    mealIDs = []
     for eachMeal in range(numMeal):
         mealID = generateMeal(planName, eachMeal)
-        g_act.linkMealIDtoPlanID(planID, mealID)
+        g_ai_dao.linkMealIDtoPlanID(planID, mealID)
 
     return
 
@@ -49,12 +49,14 @@ def generateMeal(planName: str, mealNum: int):
     calorieTotal = 500
 
     # TODO: make new meal that consists of all the foodIDs, catch error
-    mealID = g_act.createNewMealSQL(mealName, calorieTotal)
+    mealID = g_ai_dao.createNewMealSQL(mealName, calorieTotal)
 
     # TODO: link mealID with all of the foodID
     for eachFood in range(numFoods):
-        foodID = generateFood(planName)
-        g_act.linkFoodIdToMealId(mealID,foodID)
+        # TODO: REMOVE HARDCODED VALUES
+        # foodID = generateFood(planName)
+        foodID = eachFood + 1
+        g_ai_dao.linkFoodIdToMealId(mealID,foodID)
 
     return mealID
 
@@ -75,4 +77,4 @@ def generateFood(planName: str):
     # .....
 
     
-    return 5;
+    return 1;
