+++ templates/home.html
@@ -82,6 +82,7 @@
                 </div>
                 <div class="col-4">
                     <div class="card panel">
+                        {% if planID %}
                         <div class="card panel-item" style="min-height: 7.5vh;">
                             <div class="card-body" style="padding:0px">
                                 <a class="btn btn-primary" href="login.html" role="button"
@@ -117,14 +118,16 @@
                             <p class="calories">{{meal['meal_calories']}}</p>
                             <a href="/details/planid:{{planID}}/mealid:{{meal['meal_id']}}" class="btn panel-btn btn-primary"></a>
 <!-- TODO: THIS SHOULD DELETE A MEAL!!!! -->
-                            <td><a href="/TODO" class="btn btn-danger">Delete</a></td>
+                            <td><a href="/remove/planid:{{planID}}/mealid:{{meal['meal_id']}}" class="btn btn-danger">Delete</a></td>
                         </div>
                         <br>
                     {% endfor %}
+                    {% endif %}
                     </div>
                 </div>
                 <div class="col-4">
                     <div class="card panel">
+                        {% if mealID %}
                         <div class="card panel-item" style="min-height: 7.5vh;">
                             <div class="card-body" style="padding:0px">
                                 <a class="btn btn-primary" href="login.html" role="button"
@@ -139,7 +142,7 @@
                         <div class="collapse" id="addCustomFoodCollapse">
                             <div class="card card-body panel-item" style="min-height:7.5vh;">
 <!-- TODO: change the action -->
-                                <form method="post" action="/TODO">
+                                <form method="post" action="/add/food/mealid:mealID">
                                     <div class="form-group input-group">
                                         <div class="input-group-prepend">
                                             <span class="input-group-text" id="basic-addon1">Meal Name</span>
@@ -165,12 +168,13 @@
                                 <div class="card panel-item">
                                     <p>{{food['food_name']}}</p>
                                     <p class="calories">{{food['food_calories']}} cal</p>
-                                    <td><a href="/food/delete/{{food['food_id']}}" class="btn btn-danger">Delete</a></td>
+                                    <td><a href="/food/removeFromMeal/planid:{{planID}}/mealid:{{mealID}}/foodid:{{food['food_id']}}" class="btn btn-danger">Delete</a></td>
                                 </div>
                                 <br>
 
                         </tr>
                         {% endfor %}
+                        {% endif %}
                     </div>
                 </div>
             </div>
